migrate constant names check to new format
