+++ src/check/checks/mod.rs
@@ -1,3 +1,4 @@
+pub mod constant_names;
 pub mod formatting;
 pub mod script_one_pubic_run_method;
 pub mod src_names_internal;
